1. solve the memory type rotate bug; 2. update makefile support __PERF ENGINE parameter

+++ include/slice.h
@@ -95,7 +95,7 @@ inline bool operator!=(const Slice& x, const Slice& y) {
 }
 
 inline int Slice::compare(const Slice& b) const {
-    const int min_len = (size_ < b.size_) ? size_ : b.size_;
+    const int min_len = (size_ < b.size_) ? static_cast<int>(size_) : static_cast<int>(b.size_);
     int r = memcmp(data_, b.data_, min_len);
     if (r == 0) {
         if (size_ < b.size_) r = -1;
